Using sockets as IPC mechanism.
Added FAKE_APPLE build option, this will allow me to test the plugin in,
for example, Linux platforms.
Also writting a fake CMIO API for testing in Linux.
Using sockets instead of other IPC platform tied mechaninsms, this way I
can reduce the hard dependencies with the system and simplifies the
mantainance, for Windows this will allow me to test to virtual camera in
Wine, for Mac it will allow me to test the plugin outside Mac.
Also, if your are reading this, consider donating donating so I can buy
a Mac and test it on real hardware, since buying one is totaly out of my
possibilities.